move panther robot driver to separate file
